# NWUTechTrends_WebAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "NWUTechTrends_WebAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Clients": {
      "get": {
        "tags": [
          "Clients"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Client"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clients"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          }
        }
      }
    },
    "/api/Clients/{id}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clients"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/JobTelemetries/all": {
      "get": {
        "tags": [
          "JobTelemetries"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/JobTelemetry"
              }
            }
          }
        }
      }
    },
    "/api/JobTelemetries/{id}": {
      "get": {
        "tags": [
          "JobTelemetries"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/JobTelemetry"
            }
          }
        }
      },
      "put": {
        "tags": [
          "JobTelemetries"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JobTelemetry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "JobTelemetries"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "JobTelemetries"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JobTelemetry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/JobTelemetries": {
      "post": {
        "tags": [
          "JobTelemetries"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JobTelemetry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/JobTelemetry"
            }
          }
        }
      }
    },
    "/api/JobTelemetries/projectId": {
      "get": {
        "tags": [
          "JobTelemetries"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "startDate",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "endDate",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/JobTelemetries/clientID": {
      "get": {
        "tags": [
          "JobTelemetries"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "clientId",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "startDate",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "endDate",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Processes": {
      "get": {
        "tags": [
          "Processes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Process"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Processes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Process"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Process"
            }
          }
        }
      }
    },
    "/api/Processes/{id}": {
      "get": {
        "tags": [
          "Processes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Process"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Processes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Process"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Processes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Project"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        }
      }
    },
    "/api/Projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Client": {
      "type": "object",
      "properties": {
        "clientId": {
          "format": "uuid",
          "type": "string"
        },
        "clientName": {
          "type": "string"
        },
        "primaryContactEmail": {
          "type": "string"
        },
        "dateOnboarded": {
          "format": "date-time",
          "type": "string"
        },
        "jobTelemetries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/JobTelemetry"
          }
        }
      }
    },
    "JobTelemetry": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "proccesId": {
          "type": "string"
        },
        "jobId": {
          "type": "string"
        },
        "queueId": {
          "type": "string"
        },
        "stepDescription": {
          "type": "string"
        },
        "humanTime": {
          "format": "int32",
          "type": "integer"
        },
        "uniqueReference": {
          "type": "string"
        },
        "uniqueReferenceType": {
          "type": "string"
        },
        "businessFunction": {
          "type": "string"
        },
        "geography": {
          "type": "string"
        },
        "excludeFromTimeSaving": {
          "type": "boolean"
        },
        "additionalInfo": {
          "type": "string"
        },
        "entryDate": {
          "format": "date-time",
          "type": "string"
        },
        "projectId": {
          "format": "uuid",
          "type": "string"
        },
        "project": {
          "$ref": "#/definitions/Project"
        }
      }
    },
    "Process": {
      "type": "object",
      "properties": {
        "processId": {
          "format": "uuid",
          "type": "string"
        },
        "processName": {
          "type": "string"
        },
        "processType": {
          "type": "string"
        },
        "isFramework": {
          "type": "boolean"
        },
        "requiresDefaultConfig": {
          "type": "boolean"
        },
        "submitter": {
          "type": "string"
        },
        "dateSubmitted": {
          "format": "date-time",
          "type": "string"
        },
        "processConfigUrl": {
          "type": "string"
        },
        "reportUrl": {
          "type": "string"
        },
        "projectId": {
          "format": "uuid",
          "type": "string"
        },
        "defaultGeography": {
          "type": "string"
        },
        "defaultBusinessFunction": {
          "type": "string"
        },
        "platform": {
          "type": "string"
        }
      }
    },
    "Project": {
      "type": "object",
      "properties": {
        "projectId": {
          "format": "uuid",
          "type": "string"
        },
        "projectName": {
          "type": "string"
        },
        "projectDescription": {
          "type": "string"
        },
        "projectCreationDate": {
          "format": "date-time",
          "type": "string"
        },
        "projectStatus": {
          "type": "string"
        },
        "clientId": {
          "format": "uuid",
          "type": "string"
        },
        "jobTelemetries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/JobTelemetry"
          }
        }
      }
    }
  }
}